[AARCH64] Change type of __align to long long
So that ILP32 is aligned to 64bits.

+++ sysdeps/aarch64/nptl/bits/semaphore.h
@@ -31,5 +31,5 @@
 typedef union
 {
   char __size[__SIZEOF_SEM_T];
-  long int __align;
+  long long int __align;
 } sem_t;
